Refactor drogon update logic in createDrogonProjectUpdater
Refactor drogon update logic to check if drogonDir exists before executing installation steps.

+++ bin/createDrogonProjectUpdater.sh
@@ -1,11 +1,12 @@
 #! /bin/sh -
 
 #  variables
-drogonDir="~/opt/drogon/";
+
 createDrogonProjectDir="~/opt/createDrogonProject";
 exeName="createDrogonProject.sh";
 binDir="${HOME}/bin";
 optDir="~/opt/";
+drogonDir="${optDir}drogon/";
 #  update createDrogonProject
 cd $createDrogonProjectDir;
 git pull;
@@ -24,18 +25,22 @@ fi
 which drogon_ctl;
 if [ $? != 1 ];
 then
+
   #if ~/opt/drogon does not exists then
-  cp $optDir/createDrogonProject/installer.sh ~/bin/installer.sh;
-  chmod u+x ~/bin/installer.sh
-  sh ~/bin/installer.sh
-  #fi
-  cd $drogonDir;
-  git pull;
-  git submodule update --init;
-  mkdir -p build;
-  cd build;
-  cmake -DCMAKE_BUILD_TYPE=Release ..;
-  make && sudo make install;
+  if [ -e ${drogonDir} ];
+  then
+    cp $optDir/createDrogonProject/installer.sh ~/bin/installer.sh;
+    chmod u+x ~/bin/installer.sh
+    sh ~/bin/installer.sh
+  else
+    cd $drogonDir;
+    git pull;
+    git submodule update --init;
+    mkdir -p build;
+    cd build;
+    cmake -DCMAKE_BUILD_TYPE=Release ..;
+    make && sudo make install;
+    fi
 fi
 
 cp "${createDrogonProjectDir}${exeName}" "${binDir}";
